Rocket Control Script
Created a script that control the rocket. Currenly I have created 2 ways to control the rocket. I used the #if UNITY_EDITOR #endif to set the mouse click code only working on the unity editor mode. While developer try to test it on andriod phone it will skip the code inside the #if UNITY_EDITOR #endif. And the developer can moving by touching the phone screen.